pyproject.toml
[tool.poetry]
name = "eversioner"
version = "0.1.0"
description = "Creates multiple versions of an exam, with randomized questions and answers."
authors = ["Karl Busse Filho <karlfilho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
python-dotenv = "^1.0.0"
pyinstaller = "^6.10.0"
pymongo = "^4.9.1"
python-docx = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
